refactor(verify): split all zksync items
