+++ crates/verify/src/etherscan/standard_json.rs
@@ -1,4 +1,4 @@
-use super::{EtherscanSourceProvider, VerifyArgs};
+use super::{EtherscanSourceProvider, EtherscanZksyncSourceProvider, VerifyArgs};
 use crate::{provider::VerificationContext, zk_provider::ZkVerificationContext};
 use eyre::{Context, Result};
 use foundry_block_explorers::verify::CodeFormat;
@@ -52,7 +52,9 @@ impl EtherscanSourceProvider for EtherscanStandardJsonSource {
         );
         Ok((source, name, CodeFormat::StandardJsonInput))
     }
+}
 
+impl EtherscanZksyncSourceProvider for EtherscanStandardJsonSource {
     fn zk_source(
         &self,
         _args: &VerifyArgs,
